Separate shortcut key speaking in a new file
