+++ tests/unit/test_speech.py
@@ -1,7 +1,7 @@
 # A part of NonVisual Desktop Access (NVDA)
 # This file is covered by the GNU General Public License.
 # See the file COPYING for more details.
-# Copyright (C) 2021-2023 NV Access Limited, Cyrille Bougot
+# Copyright (C) 2021 NV Access Limited, Cyrille Bougot
 
 """Unit tests for the speech module.
 """
@@ -13,9 +13,6 @@ from speech import (
 	_getSpellingSpeechAddCharMode,
 	_getSpellingCharAddCapNotification,
 	_getSpellingSpeechWithoutCharMode,
-)
-from speech.speech import (
-	_getKeyboardShortcutSpeech,
 )
 from speech.commands import (
 	EndUtteranceCommand,
@@ -350,122 +347,3 @@ class Test_getSpellingSpeechWithoutCharMode(unittest.TestCase):
 			beepForCapitals=False,
 		)
 		self.assertEqual(repr(list(output)), expected)
-
-
-class Test_getKeyboardShortcutSpeech(unittest.TestCase):
-
-	def test_simpleLetterKey(self):
-		"""A shortcut consisting in only one letter."""
-
-		expected = repr([
-			CharacterModeCommand(True),
-			'A',
-			CharacterModeCommand(False),
-		])
-		output = _getKeyboardShortcutSpeech(
-			keyboardShortcut='A',
-		)
-		self.assertEqual(repr(output), expected)
-
-	def test_simpleSymbolKey(self):
-		"""A shortcut consisting in only one symbol present in symbols.dic."""
-
-		expected = repr([
-			'question',
-		])
-		output = _getKeyboardShortcutSpeech(
-			keyboardShortcut='?',
-		)
-		self.assertEqual(repr(output), expected)
-
-	def test_simpleKeyName(self):
-		"""A shortcut consisting in only a key name."""
-
-		expected = repr([
-			'Space',
-		])
-		output = _getKeyboardShortcutSpeech(
-			keyboardShortcut='Space',
-		)
-		self.assertEqual(repr(output), expected)
-
-	def test_modifiersAndLetterKey(self):
-		"""A shortcut consisting in modifiers and a letter key."""
-
-		expected = repr([
-			'Ctrl',
-			'+',
-			'Shift',
-			'+',
-			CharacterModeCommand(True),
-			'A',
-			CharacterModeCommand(False),
-		])
-		output = _getKeyboardShortcutSpeech(
-			keyboardShortcut='Ctrl+Shift+A',
-		)
-		self.assertEqual(repr(output), expected)
-
-	def test_modifierAndSymbolKey(self):
-		"""A shortcut consisting in modifiers and a symbol present in symbols.dic."""
-
-		expected = repr([
-			'Ctrl',
-			'+',
-			'slash',
-		])
-		output = _getKeyboardShortcutSpeech(
-			keyboardShortcut='Ctrl+/',
-		)
-		self.assertEqual(repr(output), expected)
-
-	def test_modifierAndPlusKey(self):
-		"""A shortcut consisting in modifiers and the + (plus) key in last position."""
-
-		expected = repr([
-			'Ctrl',
-			'+',
-			'plus',
-		])
-		output = _getKeyboardShortcutSpeech(
-			keyboardShortcut='Ctrl++',
-		)
-		self.assertEqual(repr(output), expected)
-
-	def test_modifierAndPlusKeyDescription(self):
-		"""A shortcut consisting in a modifier and the description of the + (plus) key both joined with
-		a plus surrounded by spaces. (found in Windows Magnifier)
-		"""
-
-		expected = repr([
-			'Touche de logo Windows',
-			' + ',
-			'Plus (+)',
-		])
-		output = _getKeyboardShortcutSpeech(
-			keyboardShortcut='Touche de logo Windows + Plus (+)',
-		)
-		self.assertEqual(repr(output), expected)
-
-	def test_sequentialShortcutCombiningSpacesAndCommas(self):
-		"""A sequential shortcut found in ribbons (e.g. Word)."""
-
-		expected = repr([
-			'Alt',
-			', ',
-			CharacterModeCommand(True),
-			'L',
-			CharacterModeCommand(False),
-			', ',
-			CharacterModeCommand(True),
-			'M',
-			CharacterModeCommand(False),
-			' ',
-			CharacterModeCommand(True),
-			'F',
-			CharacterModeCommand(False),
-		])
-		output = _getKeyboardShortcutSpeech(
-			keyboardShortcut='Alt, L, M F',
-		)
-		self.assertEqual(repr(output), expected)
